+++ CHANGELOG.md
@@ -11,13 +11,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+## [5.52.0] - 2023-04-04
+
+### Added
+
+- Added triggerTypes to security.
+- Added resourceNamespaces to RbacApplication.
+
 ## [5.51.0] - 2023-03-28
 
 ### Added
 
 - Adds AuthenticationMethodConfiguration derived types instances
 - Added AuthenticationMethodsPolicyMigrationState
-
 
 ## [5.50.0] - 2023-03-21